+++ retail/agents/views.py
@@ -3,7 +3,7 @@ from uuid import UUID
 
 from rest_framework import status
 from rest_framework.exceptions import NotFound, ValidationError
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -11,13 +11,24 @@ from rest_framework.viewsets import ViewSet
 
 from retail.agents.models import Agent
 from retail.agents.permissions import IsAgentOficialOrFromProjet
-from retail.agents.serializers import (PushAgentSerializer,
-                                       ReadAgentSerializer,
-                                       ReadIntegratedAgentSerializer)
+from retail.agents.serializers import (
+    AgentWebhookSerializer,
+    PushAgentSerializer,
+    ReadAgentSerializer,
+    ReadIntegratedAgentSerializer,
+)
 from retail.agents.tasks import validate_pre_approved_templates
-from retail.agents.usecases import (AssignAgentUseCase, ListAgentsUseCase,
-                                    PushAgentData, PushAgentUseCase,
-                                    RetrieveAgentUseCase)
+from retail.agents.usecases import (
+    AgentWebhookData,
+    AgentWebhookUseCase,
+    AssignAgentUseCase,
+    ListAgentsUseCase,
+    PushAgentData,
+    PushAgentUseCase,
+    RetrieveAgentUseCase,
+    UnassignAgentUseCase,
+)
+from retail.interfaces.clients.aws_lambda.client import RequestData
 from retail.internal.permissions import CanCommunicateInternally
 
 
@@ -30,19 +41,21 @@ def get_project_uuid_from_request(request: Request) -> str:
 
 
 class PushAgentView(APIView):
-    permission_classes = [CanCommunicateInternally]
-    
+    permission_classes = [IsAuthenticated]
+
     def __parse_credentials(self, agent: dict) -> list[dict]:
         credentials = []
 
         for key, credential in agent.get("credentials", {}).items():
-            credentials.append({
-                "key": key,
-                "value": credential.get("credentials", []),
-                "label": credential.get("label"),
-                "placeholder": credential.get("placeholder"),
-                "is_confidential": credential.get("is_confidential"),
-            })
+            credentials.append(
+                {
+                    "key": key,
+                    "value": credential.get("credentials", []),
+                    "label": credential.get("label"),
+                    "placeholder": credential.get("placeholder"),
+                    "is_confidential": credential.get("is_confidential"),
+                }
+            )
 
         return credentials
 
@@ -63,7 +76,7 @@ class PushAgentView(APIView):
         request_serializer.is_valid(raise_exception=True)
 
         serialized_data: PushAgentData = request_serializer.data
- 
+
         use_case = PushAgentUseCase()
         agents = use_case.execute(payload=serialized_data, files=request.FILES)
 
@@ -105,14 +118,16 @@ class AgentViewSet(ViewSet):
         return Response(response_serializer.data, status=status.HTTP_200_OK)
 
 
-class AssignAgentView(APIView):
-    permission_classes = [IsAuthenticated, IsAgentOficialOrFromProjet]
-
-    def __get_agent(self, agent_uuid: UUID) -> Agent:
+class GenericIntegratedAgentView(APIView):
+    def get_agent(self, agent_uuid: UUID) -> Agent:
         try:
             return Agent.objects.get(uuid=agent_uuid)
         except Agent.DoesNotExist:
             raise NotFound(f"Agent not found: {agent_uuid}")
+
+
+class AssignAgentView(GenericIntegratedAgentView):
+    permission_classes = [IsAuthenticated, IsAgentOficialOrFromProjet]
 
     def post(self, request: Request, agent_uuid: UUID) -> Response:
         """
@@ -128,19 +143,62 @@ class AssignAgentView(APIView):
         """
 
         project_uuid = get_project_uuid_from_request(request)
-        agent = self.__get_agent(agent_uuid)
         credentials = request.data.get("credentials", {})
+        app_uuid = request.query_params.get("app_uuid")
+
+        if app_uuid is None:
+            raise ValidationError({"app_uuid": "Missing app_uuid in params."})
+
+        agent = self.get_agent(agent_uuid)
 
         self.check_object_permissions(request, agent)
 
         use_case = AssignAgentUseCase()
-        integrated_agent, raw_client_secret = use_case.execute(agent, project_uuid, credentials)
+        integrated_agent = use_case.execute(agent, project_uuid, app_uuid, credentials)
 
-        response_serializer = ReadIntegratedAgentSerializer(
-            integrated_agent, show_client_secret=True
+        response_serializer = ReadIntegratedAgentSerializer(integrated_agent)
+
+        return Response(response_serializer.data, status=status.HTTP_201_CREATED)
+
+
+class UnassignAgentView(GenericIntegratedAgentView):
+    permission_classes = [IsAuthenticated, IsAgentOficialOrFromProjet]
+
+    def post(self, request: Request, agent_uuid: UUID) -> Response:
+        project_uuid = get_project_uuid_from_request(request)
+        agent = self.get_agent(agent_uuid)
+
+        self.check_object_permissions(request, agent)
+
+        use_case = UnassignAgentUseCase()
+        use_case.execute(agent, project_uuid)
+
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class AgentWebhookView(APIView):
+    permission_classes = [AllowAny]
+
+    def _get_data_from_request(self, request: Request) -> RequestData:
+        request_params = request.query_params
+
+        return RequestData(
+            params=request_params,
+            payload=request.data,
+            credentials={},  # TODO: Set credentials in usecase
         )
 
-        data = dict(response_serializer.data)
-        data["client_secret"] = raw_client_secret
+    def post(self, request: Request, webhook_uuid: UUID, *args, **kwargs) -> Response:
+        data = AgentWebhookData(
+            webhook_uuid=webhook_uuid,
+        )
 
-        return Response(data, status=status.HTTP_201_CREATED)
+        request_serializer = AgentWebhookSerializer(data=data)
+        request_serializer.is_valid(raise_exception=True)
+
+        request_data = self._get_data_from_request(request)
+
+        use_case = AgentWebhookUseCase()
+        lambda_return = use_case.execute(request_serializer.data, request_data)
+
+        return Response(lambda_return, status=status.HTTP_200_OK)