+++ retail/interfaces/clients/integrations/interface.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from typing import Optional
+from typing import Optional, Dict, Any
 
 
 class IntegrationsClientInterface(ABC):
@@ -33,4 +33,10 @@ class IntegrationsClientInterface(ABC):
 
     @abstractmethod
     def get_synchronized_templates(self, app_uuid: str, template_list: list) -> str:
+        pass
+
+    @abstractmethod
+    def create_library_template(
+        self, app_uuid: str, project_uuid: str, template_data: Dict[str, Any]
+    ) -> str:
         pass